refactor: extract transform expr json helpers
